feat(auth): Implementa verificação de autorização para acesso a rotas protegidas
